+++ R/entropy_overall.R
@@ -1,11 +1,18 @@
 #' Compares the observed and expected information content of the dataset
 #'
-#' Establishes how'ordered' the data is: values close to 0 indicate that combinations are highly repetitive and predictable, while values close to 1 indicate that combinations are equiprobable and prediction of future combinations is difficult
+#' Establishes how'ordered' the data is: values close to 0 indicate that
+#' combinations are highly repetitive and predictable, while values close to 1
+#' indicate that combinations are equiprobable and prediction of future
+#' combinations is difficult
 #'
 #'
 #' @param netfacs.data object resulting from netfacs() function
 #'
-#' @return Function returns the ratio of observed entropy/expected entropy. Expected entropy is based on randomization (shuffling the observed elements while maintaining the number of elements per row) and represents the maximum entropy a dataset with the same properties as this one can reach. Ratios closer to 0 are more ordered; ratios closer to 1 are more random.
+#' @return Function returns the ratio of observed entropy/expected entropy.
+#'   Expected entropy is based on randomization (shuffling the observed elements
+#'   while maintaining the number of elements per row) and represents the
+#'   maximum entropy a dataset with the same properties as this one can reach.
+#'   Ratios closer to 0 are more ordered; ratios closer to 1 are more random.
 #' @importFrom picante randomizeMatrix
 #' @export
 #'