+++ tests/testthat/test_netfacs.R
@@ -1,4 +1,4 @@
-test_that("single core, condition specified is correct",{
+test_that("condition specified calculations are correct",{
   expect_equal(netfacs(d.sim.with.context, 
                        condition = rownames(d.sim.with.context), 
                        test.condition = "a", 
@@ -8,8 +8,6 @@ test_that("single core, condition specified is correct",{
                        n_cores = 1),
                res.single.core.with.context, 
                tolerance = 0.13)
-})
-test_that("multi-core, condition specified is correct",{
   expect_equal(netfacs(d.sim.with.context, 
                        condition = rownames(d.sim.with.context), 
                        test.condition = "a", 
@@ -20,7 +18,8 @@ test_that("multi-core, condition specified is correct",{
                res.multi.core.with.context, 
                tolerance = 0.13)
 })
-test_that("single core, condition NOT specified is correct",{
+test_that("condition NOT specified calculations are correct",{
+  # single core
   expect_equal(netfacs(d.sim.no.context, 
                        ran.trials = 500,
                        combination.size = 2,
@@ -28,8 +27,7 @@ test_that("single core, condition NOT specified is correct",{
                        n_cores = 1),
                res.single.core.no.context, 
                tolerance = 0.11)
-})
-test_that("multi-core, condition NOT specified is correct",{
+  # multi core
   expect_equal(netfacs(d.sim.no.context, 
                        ran.trials = 500, 
                        combination.size = 2,
@@ -38,55 +36,66 @@ test_that("multi-core, condition NOT specified is correct",{
                res.multi.core.no.context, 
                tolerance = 0.11)
 })
-test_that("error message is given when data has NA",{
-  expect_error(
-    {
-      netfacs(
-        data = NA,
-        condition = emotions_set[[2]]$emotion,
-        test.condition = "anger",
-        null.condition = NULL,
-        ran.trials = 20
-      )
-    }
-  )
+test_that("error message is given when data has NAs",{
+  expect_error({
+    netfacs(
+      data = NA,
+      condition = emotions_set[[2]]$emotion,
+      test.condition = "anger",
+      null.condition = NULL,
+      ran.trials = 20)
+  })
+  expect_error({
+    netfacs(
+      data = emotions_set[[1]],
+      condition = NA,
+      test.condition = "anger",
+      null.condition = NULL,
+      ran.trials = 20)
+  })
 })
-test_that("error message is given when condidion has NA",{
-  expect_error(
-    {
-      netfacs(
-        data = emotions_set[[1]],
-        condition = NA,
-        test.condition = "anger",
-        null.condition = NULL,
-        ran.trials = 20
-      )
-    }
-  )
+test_that("error message is given when condidions are not misspecified",{
+  expect_error({
+    netfacs(
+      data = emotions_set[[1]],
+      condition = emotions_set[[2]]$emotion,
+      test.condition = "AAA",
+      null.condition = "anger",
+      ran.trials = 20)
+  })
+  expect_error({
+    netfacs(
+      data = emotions_set[[1]],
+      condition = emotions_set[[2]]$emotion,
+      test.condition = "anger",
+      null.condition = "AAA",
+      ran.trials = 20)
+  })
+  expect_error({
+    netfacs(
+      data = emotions_set[[1]],
+      condition = "b",
+      test.condition = "b",
+      null.condition = NULL,
+      ran.trials = 10)
+  })
 })
-test_that("error message is given when test.condidion is not part of condition",{
-  expect_error(
-    {
-      netfacs(
-        data = emotions_set[[1]],
-        condition = emotions_set[[2]]$emotion,
-        test.condition = "AAA",
-        null.condition = "anger",
-        ran.trials = 20
-      )
-    }
-  )
+test_that("warning message is given when test or null conditions are specified with a NULL condition vector",{
+  expect_warning({
+    netfacs(
+      data = emotions_set[[1]],
+      condition = NULL,
+      test.condition = "b",
+      null.condition = NULL,
+      ran.trials = 10)
+  })
+  expect_warning({
+    netfacs(
+      data = emotions_set[[1]],
+      condition = NULL,
+      test.condition = NULL,
+      null.condition = "a",
+      ran.trials = 10)
+  })
 })
-test_that("error message is given when null.condidion is not part of condition",{
-  expect_error(
-    {
-      netfacs(
-        data = emotions_set[[1]],
-        condition = emotions_set[[2]]$emotion,
-        test.condition = "anger",
-        null.condition = "AAA",
-        ran.trials = 20
-      )
-    }
-  )
-})
+